requirements.txt
python-pptx
pillow
pytesseract
tqdm
regex
python-dateutil
rapidfuzz
python-dotenv
google-generativeai
